refactor: rewrite Panel
