+++ components/Panel/Panel.tsx
@@ -1,8 +1,6 @@
 "use client";
 
-import { Transition } from "@headlessui/react";
-import * as DialogPrimitive from "@radix-ui/react-dialog";
-import { Fragment, Suspense } from "react";
+import { Suspense } from "react";
 import { useAtom, useAtomValue, useSetAtom } from "jotai/react";
 import {
   currentPanelAtom,
@@ -15,6 +13,7 @@ import { EPListContent } from "@/components/Panel/EPListContent";
 import { PersonListContent } from "@/components/Panel/PersonListContent";
 import { SubjectContentSkeleton } from "@/components/Skeleton/SubjectContentSkeleton";
 import { RelationSubjectContent } from "@/components/Panel/RelationSubjectContent";
+import { Dialog, DialogContent_Panel } from "@/ui/primitive/Dialog";
 
 export function Panel() {
   const panel = useAtomValue(currentPanelAtom);
@@ -22,7 +21,7 @@ export function Panel() {
   const [isOpenPanel, setIsOpenPanel] = useAtom(isOpenPanelAtom);
 
   return (
-    <DialogPrimitive.Root
+    <Dialog
       open={isOpenPanel}
       onOpenChange={(open) => {
         if (!open) {
@@ -33,71 +32,21 @@ export function Panel() {
         }
       }}
     >
-      <DialogPrimitive.Portal forceMount>
-        <Transition.Root show={isOpenPanel}>
-          <Transition.Child
-            as={Fragment}
-            enter="ease-out duration-300"
-            enterFrom="opacity-0"
-            enterTo="opacity-100"
-            leave="ease-in duration-200"
-            leaveFrom="opacity-100"
-            leaveTo="opacity-0"
-          >
-            <DialogPrimitive.Overlay
-              forceMount
-              id="panel-overlay"
-              className="fixed inset-0 z-40 grid place-items-center overflow-y-auto bg-blackA-9 pb-16 dark:bg-whiteA-9"
-            >
-              <Transition.Child
-                as={Fragment}
-                enter="ease-out duration-300"
-                enterFrom="opacity-0 scale-95"
-                enterTo="opacity-100 scale-100"
-                leave="ease-in duration-200"
-                leaveFrom="opacity-100 scale-100"
-                leaveTo="opacity-0 scale-95"
-              >
-                <DialogPrimitive.Content
-                  forceMount
-                  className="z-40 mt-16 flex w-[1040.62px] flex-col rounded-lg bg-neutral-1 pt-4 pb-6 shadow-lg outline-none"
-                  onPointerDownOutside={(e) => {
-                    if (
-                      e.detail.originalEvent.which === 2 ||
-                      e.detail.originalEvent.button === 4
-                    ) {
-                      e.preventDefault();
-                    }
-
-                    // https://github.com/radix-ui/primitives/issues/1280#issuecomment-1198248523
-                    const currentTarget = e.currentTarget as HTMLElement;
-
-                    if (
-                      e.detail.originalEvent.offsetX > currentTarget.clientWidth
-                    ) {
-                      e.preventDefault();
-                    }
-                  }}
-                >
-                  {panel?.type === "subject" ? (
-                    <Suspense fallback={<SubjectContentSkeleton />}>
-                      <SubjectContent subject_id={panel.target_id} />
-                    </Suspense>
-                  ) : panel?.type === "characterList" ? (
-                    <CharacterListContent subject_id={panel.target_id} />
-                  ) : panel?.type === "episodeList" ? (
-                    <EPListContent subject_id={panel.target_id} />
-                  ) : panel?.type === "personList" ? (
-                    <PersonListContent subject_id={panel.target_id} />
-                  ) : panel?.type === "subjectList" ? (
-                    <RelationSubjectContent subject_id={panel.target_id} />
-                  ) : null}
-                </DialogPrimitive.Content>
-              </Transition.Child>
-            </DialogPrimitive.Overlay>
-          </Transition.Child>
-        </Transition.Root>
-      </DialogPrimitive.Portal>
-    </DialogPrimitive.Root>
+      <DialogContent_Panel isOpen={isOpenPanel}>
+        {panel?.type === "subject" ? (
+          <Suspense fallback={<SubjectContentSkeleton />}>
+            <SubjectContent subject_id={panel.target_id} />
+          </Suspense>
+        ) : panel?.type === "characterList" ? (
+          <CharacterListContent subject_id={panel.target_id} />
+        ) : panel?.type === "episodeList" ? (
+          <EPListContent subject_id={panel.target_id} />
+        ) : panel?.type === "personList" ? (
+          <PersonListContent subject_id={panel.target_id} />
+        ) : panel?.type === "subjectList" ? (
+          <RelationSubjectContent subject_id={panel.target_id} />
+        ) : null}
+      </DialogContent_Panel>
+    </Dialog>
   );
 }
